:truck: refactor: rename auth screen error state object structure
