+++ feature/auth/src/main/java/com/anna/greeneats/auth/common/AuthScreenErrorState.kt
@@ -1,5 +1,6 @@
-package com.anna.greeneats.auth.login.state
+package com.anna.greeneats.auth.common
 
+import com.anna.greeneats.core.util.validation.error.ConfirmPassword
 import com.anna.greeneats.core.util.validation.error.Email
 import com.anna.greeneats.core.util.validation.error.Password
 
@@ -11,4 +12,9 @@ data class  EmailErrorState(
 data class PasswordErrorState(
   val isPasswordError: Boolean = false,
   val passwordValidation: Password = Password.INITIAL
+)
+
+data class ConfirmPasswordErrorState(
+  val isPasswordError: Boolean = false,
+  val confirmPasswordValidation: ConfirmPassword = ConfirmPassword.INITIAL
 )
